+++ src/lib/Tick.sol
@@ -10,11 +10,11 @@ library Tick {
     function update(
         mapping(int24 => Tick.Info) storage self,
         int24 tick,
-        int128 liquidityDelta
+        uint128 liquidityDelta
     ) internal {
         Tick.Info storage tickInfo = self[tick];
         uint128 liquidityBefore = tickInfo.liquidity;
-        uint128 liquidityAfter = liquidityBefore + uint128(liquidityDelta);
+        uint128 liquidityAfter = liquidityBefore + liquidityDelta;
 
         if (liquidityBefore == 0) {
             tickInfo.initialized = true;
